Spruced up for current dart
DOM library now allows writing to canvas, so do that.

Add build script for building with frog compiler.

Time rendering, write stats to console as well as screen.

(Screen doesn't work on Firefox for some reason.)

Change name of "write" global method to not interfere with frog
runtime method with the same name. (Possibly a bug in frog?)